requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Config parsing
pyyaml>=6.0

# PV physics modelling
pvlib>=0.10.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Parquet I/O
pyarrow>=12.0.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0
requests>=2.31.0

# Deep learning (LSTM forecaster and residual correction)
torch>=2.0.0

# Model serialisation
joblib>=1.3.0
